refactor app logging
