+++ scout/log/__init__.py
@@ -1,2 +1,1 @@
-from .handlers import TlsSMTPHandler
 from .log import init_log
